Add reviews, with reviewer's permission
Hi Jason,

the reviewers are Ok with you putting the reviews up and appreciate that
you asked first.

There was some concern that as general policy this might not be a good
idea, or at least would need discussion by the wider community, but as
an isolated request it is not a problem.

Cheers,
Gerwin